Pin `setuptools_scm` to <8
There are some breaking changes, and this is making it so that the
`_version.py` file generated is not compatible with Python 2 anymore.
I don't see an obvious or easy way to figure out what version of
`setuptools_scm` is in use, so we'll just pin the version for now.

I recommend removing `setuptools_scm` entirely in the future.